Switched file chooser to use javaFX FileChooser
javaFX FileChooser utilizes the native Windows file chooser (unlike the
last check-in that simply mimicked (poorly) the Windows file chooser).
Leveraging a javaFX component in a non-javaFX program was quite
complicated, and I relied on a stackOverflow solution to get it working.

Also changed a bunch of static variables in ControlPanel class to
non-static.

Added getUncheckedIndices method to Palette class.