Added ability to use AppData directory
Added ability to use AppData directory as opposed to only using the
current working directory. (Saving the config file(s) on the fly when
they reside in ProgramFiles didn't work unless the program was run as
administrator.)

Also updated installer to NOT include the Java installer. Instead, asks
if user wants to go to Java.com to download and install Java.